refactor: merge utils
