+++ packages/blocks/src/page-block/edgeless/mode-controllers/connector/grid.ts
